requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
jinja2==3.1.2
python-dotenv==1.0.0
chromadb==0.4.18
faiss-cpu>=1.8.0
sentence-transformers==2.2.2
pymupdf==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
openai==1.3.5
PyPDF2==3.0.1
reportlab==4.0.7
markdown==3.5.1
requests==2.31.0
sqlalchemy==2.0.23
flask-jwt-extended==4.5.3
flask-bcrypt==1.0.1
werkzeug==3.0.1
openpyxl==3.1.2
langchain==0.0.350
langchain-openai==0.0.2
easyocr==1.7.0


